requirements.txt
beautifulsoup4
celery==5.0.2
channels==2.4.0
channels-redis==2.4.2
daphne==2.5.0
Django==3.1.3
djangorestframework==3.12.1
django-filter
django-oauth-toolkit==1.7.1
dnspython
paramiko
psycopg2-binary==2.9.3
python-decouple
python-digitalocean
pytest-django==4.5.2
ratelimiter
redis==3.5.3
requests
shortuuid
